Client polls for yield
Client can support yield by polling every second after prompting
if yield is on.

The ship options were made into a global are removed from
promptOptions. This way they can be manipulated by any code
in a callback (like the setTimeout). They only change on
running a message so that's the only time the global needs to
be set.