** Express Refactor **  Step 2: Remove all old express code
This is the second step of replacing the old server code
with a new setup based on a fork of nodebootstrap.
Some of this code may show up later, but it will be in a new file
so might as well remove it all and bring it back in when needed.

So long old code!
You won't be forgotten!