ship's commandHolders moved into shipModel and added simple views
Rather than the shipModel containing a generic container of each
commandHolder, there is a commandHolderModel per individual ship
commandHolders.

This allows commandHolderViews to be created for each specific
commandHolder needed by the ship and tied to a specific el
by the appView.

Downside is that every command must be in a commandHolder and
a new commandHolderModel and commandHolderView needs to be created
for each holder added to the ship.